requirements.txt
# AI Smart Agriculture Assistant - Requirements
# All packages are FREE and open-source
# Install with: pip install -r requirements.txt

# ── Core Framework ────────────────────────────────────────────────────────────
streamlit>=1.28.0

# ── Machine Learning (Crop Recommendation) ───────────────────────────────────
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# ── Image Processing (Disease Detection) ─────────────────────────────────────
Pillow>=10.0.0

# ── HTTP Requests (Weather API) ──────────────────────────────────────────────
requests>=2.31.0

# ── Optional: Deep Learning (Disease Detection CNN) ──────────────────────────
# Uncomment for full CNN-based disease detection (larger install ~500MB):
# tensorflow>=2.13.0
# OR for lighter installation:
# tensorflow-cpu>=2.13.0

# ── Data Visualization ────────────────────────────────────────────────────────
matplotlib>=3.7.0
plotly>=5.17.0

# ── Utility ──────────────────────────────────────────────────────────────────
python-dateutil>=2.8.2
